adding changes according to the comments

+++ 1-Programming/src/main/java/typeConversions/Exercise3.java
@@ -3,7 +3,7 @@ package typeConversions;
 public class Exercise3 {
     public static void main(String[] args) {
         double n = 8.202;
-        int rest = (int) Math.round(n);
+        double rest = Math.round(n);
         System.out.println(rest);
     }
 }
